// src/interfaces/constants.ts
export enum HOLIDAYTYPE {
  EVENTS = '#82E0AA',
  HOLIDAY = '#3498DB ',
  OPTIONAL_HOLIDAY = '#A569BD',
  BIRTHDAY = '#E74C3C',
  WORK_ANNIVERSARIES = '#0B5345',
  OTHER = '#4A235A'
}

export enum USERROLE {
  SUPERADMIN = '1001',
  HR = '4321',
  ACCOUNTANT = '2222',
  MANAGER = '4327',
  ADMIN = '3462',
  EMPLOYEE = '2986',
  COMPANY = '3435',
  COMPANYADMIN = '3850',
}